little bit of spit and polish for the image.  added eye of gnome for doing the slideshow, leaving xiv in the image for fun, since the time was taken to get the recipe to work.  changed the sync_and_patch script to only do things if the folders exist, some sanity checking

+++ recipes/cfa-prebuilt/cfa-prebuilt/S30slideshow-link.sh
@@ -18,7 +18,9 @@ for x in $CMDLINE; do
             echo creating slideshow symlink | tee /dev/kmsg > /dev/null
         fi
         echo "270" > /etc/rotation
+        if [ -e /home/root/.configuration/xorg.conf_cfa10049 ]; then
         mv /home/root/.configuration/xorg.conf_cfa10049 /etc/X11/xorg.conf
+		fi
         exit 0;
         ;;
     CFA-10055)
@@ -27,7 +29,9 @@ for x in $CMDLINE; do
             echo creating slideshow symlink | tee /dev/kmsg > /dev/null
         fi
         echo "270" > /etc/rotation
+        if [ -e /home/root/.configuration/xorg.conf_cfa10055 ]; then
         mv /home/root/.configuration/xorg.conf_cfa10055 /etc/X11/xorg.conf
+		fi
         exit 0;
         ;;
     CFA-10056)
@@ -36,7 +40,9 @@ for x in $CMDLINE; do
             echo creating slideshow symlink | tee /dev/kmsg > /dev/null
         fi
         echo "270" > /etc/rotation
+        if [ -e /home/root/.configuration/xorg.conf_cfa10056 ]; then
         mv /home/root/.configuration/xorg.conf_cfa10056 /etc/X11/xorg.conf
+        fi
         exit 0;
         ;;
     CFA-10057)
@@ -45,7 +51,9 @@ for x in $CMDLINE; do
             echo creating slideshow symlink | tee /dev/kmsg > /dev/null
         fi
         echo "90" > /etc/rotation
+        if [ -e /home/root/.configuration/xorg.conf_cfa10057 ]; then
         mv /home/root/.configuration/xorg.conf_cfa10057 /etc/X11/xorg.conf
+		fi
         exit 0;
         ;;
     CFA-10058)
@@ -53,7 +61,9 @@ for x in $CMDLINE; do
             busybox ln -s /home/root/images/CFA10058_slideshow /home/root/images/slideshow.sh
             echo creating slideshow symlink | tee /dev/kmsg > /dev/null
         fi
+        if [ -e /home/root/.configuration/xorg.conf_cfa10058 ]; then
         mv /home/root/.configuration/xorg.conf_cfa10058 /etc/X11/xorg.conf
+		fi
         exit 0;
         ;;
     esac
